Implement the signal-sending helper.
The implementation uses a couple of Linux extensions to POSIX,
pipe2 and ppoll.  These could be avoided by using a condition variable
instead of a pipe to tell the signaler thread to exit, but I don’t
believe in condition variables.  AFAICT anything you can do with a
condition variable can be done more efficiently, more flexibly, and
with less ceremony by using either a semaphore or a pipe.  On the
other hand, the most natural way to use a pipe for this particular
job does require use of these extensions.