Convert examples/signaler_usage.py to a proper set of unit tests.
One test takes much longer than it ought to, despite succeeding.
This is because signaler.c is sending signals specifically to a
secondary Python thread.  The C-level signal handler runs on that
thread, but all it does is set a flag telling the interpreter that
it needs to run Python signal handlers Real Soon Now.  That flag
is only ever checked on the _initial_ Python thread, and that thread
is blocked on a system call, which doesn’t get interrupted by the
signal, because the signal was handled on a different thread.
This will be addressed in the next commit.