Let’s actually take some measurements!
Implement a latency benchmark, using numpy.fft.fft as our complex
algorithm with a long runtime.  The files in measurements/ demonstrate
that numpy’s FFT has *not* been coded to be interruptible; its median
runtime on an input of 2**24 random samples is 0.9328139 seconds on my
computer, with very little variation (IQR 0.0031326 s).  If a
KeyboardInterrupt is sent 5 ms after the computation begins, the FFT
still runs to completion, taking almost exactly the same time.