Put the syntax highlighting back in. By hand.
Well, with a little help from Pygments, but a lot of manual tweaking.

The nice thing about doing it this way is we no longer need most of
the stacks, we can put fragment markers right into the code.

And the other nice thing about doing it this way is the reveal.js
syntax highlighting plugin’s bugs are no longer of any concern to me.